refactor: moving components
